+++ stoilib/src/constants.rs
@@ -10,7 +10,7 @@ pub const HALF_FRAME: usize = FRAME_LENGTH / 2;
 pub const HOP_LENGTH: usize = FRAME_LENGTH / 2;
 
 // Internal sampling frequency for STOI computation
-pub const FS: u32 = 10_000;
+pub const FS: usize = 10_000;
 
 pub const FFT_LENGTH: usize = 512;
 pub const FFT_BINS: usize = FFT_LENGTH / 2 + 1;
